refactor: split orchestrator commands into submodules
