+++ src-tauri/src/commands/orchestrator/mod.rs
@@ -1,0 +1,25 @@
+mod actions;
+mod sessions;
+mod stream;
+mod stream_api;
+mod stream_cli;
+mod types;
+
+pub use actions::{__cmd__process_orchestrator_response, __cmd__set_orchestrator_error};
+pub use actions::{process_orchestrator_response, set_orchestrator_error};
+pub use sessions::{
+    __cmd__clear_chat_history, __cmd__create_chat_session, __cmd__delete_chat_session,
+    __cmd__get_active_chat_session, __cmd__get_chat_history, __cmd__get_orchestrator_context,
+    __cmd__get_orchestrator_session, __cmd__list_chat_sessions, __cmd__reset_cli_session,
+};
+pub use sessions::{
+    clear_chat_history, create_chat_session, delete_chat_session, get_active_chat_session,
+    get_chat_history, get_orchestrator_context, get_orchestrator_session, list_chat_sessions,
+    reset_cli_session,
+};
+pub use stream::{__cmd__cancel_orchestrator_chat, __cmd__stream_orchestrator_chat};
+pub use stream::{cancel_orchestrator_chat, stream_orchestrator_chat};
+pub use types::{
+    ApiStreamRegistry, OrchestratorAction, OrchestratorContext, OrchestratorEvent,
+    OrchestratorResponse,
+};
